requirements.txt
asgiref==3.8.1
azure-core==1.34.0
azure-messaging-webpubsubservice==1.2.2
certifi==2025.4.26
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
Django==5.2.1
idna==3.10
isodate==0.7.2
msgpack==1.1.1
pillow==11.2.1
PyJWT==2.10.1
redis==6.2.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
mysqlclient>=2.1
